drm/amd/display: Move bw_fixed macros to header file
The macros defined at bw_fixed are important mathematical definitions,
specifying masks to get the fractional part and the maximum and
minimum values of I64. In order to enable unit tests for bw_fixed,
it is relevant to have access to those macros. This commit moves the
macros to the header file, making it accessible to future unit tests.

Co-developed-by: Magali Lemes <magalilemes00@gmail.com>
Signed-off-by: Magali Lemes <magalilemes00@gmail.com>
Signed-off-by: Maíra Canal <maira.canal@usp.br>
Co-developed-by: Tales Aparecida <tales.aparecida@gmail.com>
Signed-off-by: Tales Aparecida <tales.aparecida@gmail.com>